Evaluate models on custom dataset

+++ utils.py
@@ -135,7 +135,7 @@ def load_custom_test_data(path, channels=3, use_sparse_depth=False, dont_interpo
                 img = np.stack([img[:,:,0], img[:,:,1], img[:,:,2], iz, vm], axis=-1)
             else:
                 img = np.stack([img[:,:,0], img[:,:,1], img[:,:,2], iz], axis=-1)
-        depth = np.asarray(Image.open( rgb_path.replace('interp_depth', 'ground_truth') ))/gt_divider # convert from mm to m
+        depth = np.asarray(Image.open( rgb_path.replace('interp_depth', 'ground_truth') ), dtype=np.float32)/gt_divider # convert from mm to m
         images.append(img)
         depths.append(depth)
